delete the two config fields nothing reads, and test for the next one
`precision.reduce_dtype` had no effect on training. Setting it to bfloat16
against float32 left the parameters bit-identical on all four paths -- single
process, FSDP, dp=2 gradient reduction and tp=2 row-projection all-reduce, which
is the one its docstring claimed to reach. Every one of those produced DTensor
parameters, so the sharding was real; only the dtype knob did nothing.

A sensitivity control says that null result is a real absence and not a blind
test: wiring bf16 into FSDP's MixedPrecisionPolicy by hand moved the result by
1.69e-02.

It had not always been dead. Before the DTensor migration,
`RowParallelLinear(reduce_dtype=...)` passed it to `reduce_from_tp`, and the
migration deleted that layer in favour of a `RowwiseParallel` style that takes no
dtype. The connection went; everything about it stayed -- the field, its
validation, two presets that set it, an example config, and a docstring asserting
"threaded into the TP row projections". Validation made it look more real than a
live field: an invalid value still raises.

So a user on ConfigPreset.performance() believed they had bf16 reductions, saw
'reduce_dtype': 'bfloat16' in every dump, and trained in fp32 throughout.

An AST scan for readers (`.field` or getattr outside schema.py; a field's own
type check does not count, being what a dead field has left) left exactly one
more: `overlap.enable_parameter_prefetch`. It is deader still -- the other
advanced overlap switches at least reach capability.py's rejection guard, and
this one does not even do that. Deleted alongside.

`test_every_config_field_is_read_somewhere` now refuses the next one. This defect
class has appeared at least seven times and has a property no existing test could
see: nothing fails when a field stops being read. Three controls: re-adding
reduce_dtype fails it, a brand-new unread field fails it, and a field that IS
read does not.

254 passed (+1), ruff 90 unchanged, 3 contracts kept, all 7 examples and the CLI
validate clean.

+++ src/aitrainer/presets.py
@@ -41,7 +41,7 @@ class ConfigPreset:
         """
         return FrameworkConfig(precision=PrecisionConfig(
             compute_dtype="bfloat16", grad_dtype="float32",
-            reduce_dtype="bfloat16", optimizer_dtype="bfloat16"))
+            optimizer_dtype="bfloat16"))
 
     @staticmethod
     def performance() -> FrameworkConfig:
@@ -53,4 +53,4 @@ class ConfigPreset:
         """
         return FrameworkConfig(precision=PrecisionConfig(
             compute_dtype="bfloat16", grad_dtype="float32",
-            reduce_dtype="bfloat16", optimizer_dtype="float32"))
+            optimizer_dtype="float32"))
